refactor: move utils to metricsExplorer util
